refactor: migrate purchased_tile_riches tests to scenario tables (#3856)
Collapse 12 imperative test bodies into purchased_tile_riches_scenarios.dart
with shared post-conquest/tribe/multi-GP builders; economy test LOC now ≤6,500.
